+++ contracts/finality/src/exec/finality.rs
@@ -1,14 +1,15 @@
 use crate::custom_queries::get_current_epoch;
 use crate::error::ContractError;
 use crate::msg::BabylonMsg;
-use crate::signing_context::{get_fp_fin_vote_context_v0, get_fp_rand_commit_context_v0};
 use crate::state::config::CONFIG;
 use crate::state::finality::{get_finality_signature, insert_finality_sig_and_signatory};
 use crate::state::public_randomness::{
     get_timestamped_pub_rand_commit_for_height, insert_pub_rand_commit, insert_pub_rand_value,
     PubRandCommit,
 };
-use crate::utils::query_finality_provider;
+use crate::utils::{
+    get_fp_fin_vote_context_v0, get_fp_rand_commit_context_v0, query_finality_provider,
+};
 use babylon_bindings::BabylonQuery;
 use babylon_merkle::Proof;
 use cosmwasm_std::{Deps, DepsMut, Env, Event, Response};
