+++ contracts/finality/src/lib.rs
@@ -12,7 +12,6 @@ pub mod error;
 pub mod exec;
 pub mod msg;
 pub mod queries;
-pub mod signing_context;
 pub mod state;
 pub mod utils;
